refactor!: separate "load" into its own directory
