+++ raillabel/__init__.py
@@ -7,7 +7,7 @@ from . import _util, format
 from .exceptions import *
 from .filter.filter import filter
 from .format.scene import Scene
-from .load import load
+from .load_.load import load
 from .save import save
 from .validate.validate import validate
