+++ raillabel/load_/__init__.py
@@ -1,0 +1,3 @@
+# Copyright DB Netz AG and contributors
+# SPDX-License-Identifier: Apache-2.0
+"""Package containing all necessary modules for loading."""
